src/requirements.txt
Flask==0.11.1
pytest==2.9.1
PyYAML==3.11
flake8==2.5.4
flask-restplus==0.9.2
raven[flask]==5.29.0
SQLAlchemy==1.1.2
alembic==0.8.8
psycopg2==2.6.2
SQLAlchemy-Utils==0.32.9
pycountry==1.20
pytest-flask==0.10.0
python-json-logger==0.1.5
